22-03-23

+++ forum/src/main/java/com/projet/forum/Services/MessageServices/MessageServiceImpl.java
@@ -38,18 +38,25 @@ public class MessageServiceImpl implements MessageService{
         repository.save(message);
     }*/
 
-    @Override public void modifyMessage(Long id, String text){
+    @Override public void modifyMessage(Long uId, Long id, String text){
 
+        UserEntity user = u_repository.findById(uId).orElseThrow();
         MessageEntity message = repository.findById(id).orElseThrow();
-        
-        if(!(message.getContent().equals(text)))
-            message.setContent(text);
-        else throw new EqualMessageException("You can't change message to the same text");
+
+        if(user.getId().equals(message.getUser_author().getId()) || user.getRole().equals(Role.ADMIN)){
+            
+            if(!(message.getContent().equals(text)))
+                message.setContent(text);
+            else throw new EqualMessageException("You can't change message to the same text");
+        }
     }
 
-    @Override public void deleteMessage(Long id){
+    @Override public void deleteMessage(Long uId, Long id){
 
+        UserEntity user = u_repository.findById(uId).orElseThrow();
         MessageEntity message = repository.findById(id).orElseThrow();
+
+        if(user.getId().equals(message.getUser_author().getId()) || user.getRole().equals(Role.ADMIN))
         repository.delete(message);
     }
 }
